requirements.txt
python-telegram-bot==20.7
solders==0.20.0
solana==0.34.0
anchorpy==0.20.0
aiohttp==3.9.3
aiosqlite==0.20.0
cryptography==42.0.5
base58==2.1.1
construct==2.10.68
borsh-construct==0.1.0
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.1
apscheduler==3.10.4
cachetools==5.3.3
pydantic==2.6.3
loguru==0.7.2
tenacity==8.2.3
